Revert "feat(fxa-payments-server): add paypal to subscription management"
This reverts commit 1379ec958fe01e247462da6dcdd6f6ef7186fac5.
