+++ packages/fxa-auth-server/lib/routes/subscriptions/stripe.ts
@@ -48,10 +48,7 @@ export type PaypalPaymentError = 'missing_agreement' | 'funding_source';
 
 type PaymentBillingDetails = ReturnType<
   StripeHandler['extractBillingDetails']
-> & {
-  paypal_payment_error?: PaypalPaymentError;
-  billing_agreement_id?: string;
-};
+> & { paypal_payment_error?: PaypalPaymentError };
 
 export class StripeHandler {
   constructor(
@@ -374,12 +371,6 @@ export class StripeHandler {
     const billingDetails = this.extractBillingDetails(
       customer
     ) as PaymentBillingDetails;
-
-    if (billingDetails.payment_provider === 'paypal') {
-      billingDetails.billing_agreement_id = this.stripeHelper.getCustomerPaypalAgreement(
-        customer
-      );
-    }
 
     if (
       billingDetails.payment_provider === 'paypal' &&
