+++ packages/fxa-payments-server/server/lib/server.js
@@ -88,7 +88,6 @@ module.exports = () => {
       },
     },
     paypal: {
-      apiUrl: config.get('paypal.apiUrl'),
       clientId: config.get('paypal.clientId'),
       scriptUrl: config.get('paypal.scriptUrl'),
     },
