+++ packages/fxa-payments-server/src/components/AlertBar/index.scss
@@ -17,7 +17,6 @@
   text-align: center;
 }
 
-#top-bar .alertError,
 #top-bar .alertSuccess {
   background: #1d1133;
   color: #fff;
@@ -32,10 +31,6 @@
       fill: #fff;
     }
   }
-}
-
-#top-bar .alertError {
-  background: #FF4F5E;
 }
 
 .alert span.checked::before {
