+++ packages/fxa-payments-server/src/components/DialogMessage/index.scss
@@ -3,7 +3,6 @@
 
 #dialogs {
   .modal {
-    border-radius: 12px;
     padding: 25px 27px 30px;
     text-align: center;
     width: 545px;
