+++ packages/fxa-payments-server/src/lib/apiClient.ts
@@ -374,33 +374,3 @@ export async function apiDetachFailedPaymentMethod(params: {
   );
   return result;
 }
-
-export async function apiUpdateBillingAgreement(params: {
-  token: string;
-}): Promise<{
-  invoice_settings: Object;
-  subscriptions: Subscription[];
-}> {
-  const metricsOptions: Amplitude.EventProperties = {
-    paymentProvider: 'paypal',
-  };
-  Amplitude.updateDefaultPaymentMethod_PENDING(metricsOptions);
-  try {
-    const response = await apiFetch(
-      'POST',
-      `${config.servers.auth.url}/v1/oauth/subscriptions/paymentmethod/billing-agreement`,
-      { body: JSON.stringify(params) }
-    );
-    Amplitude.updateDefaultPaymentMethod_FULFILLED({
-      ...metricsOptions,
-    });
-
-    return response;
-  } catch (error) {
-    Amplitude.updateDefaultPaymentMethod_REJECTED({
-      ...metricsOptions,
-      error,
-    });
-    throw error;
-  }
-}
