+++ packages/fxa-payments-server/src/lib/config.test.ts
@@ -144,7 +144,6 @@ const expectedMergedConfig = {
     },
   },
   paypal: {
-    apiUrl: '',
     clientId: '',
     scriptUrl: '',
   },
