+++ packages/fxa-payments-server/src/lib/config.ts
@@ -33,7 +33,6 @@ export interface Config {
     };
   };
   paypal: {
-    apiUrl: string;
     clientId: string;
     scriptUrl: string;
   };
@@ -77,7 +76,6 @@ export function defaultConfig(): Config {
       },
     },
     paypal: {
-      apiUrl: '',
       clientId: '',
       scriptUrl: '',
     },
