+++ packages/fxa-payments-server/src/lib/hooks.tsx
@@ -2,15 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import React, {
-  useCallback,
-  useState,
-  useEffect,
-  useRef,
-  ChangeEvent,
-} from 'react';
+import { useCallback, useState, useEffect, useRef, ChangeEvent } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import { ButtonBaseProps } from '../components/PayPalButton';
 
 export function useCallbackOnce(cb: Function, deps: any[]) {
   const called = useRef(false);
@@ -61,44 +54,4 @@ export function useMatchMedia(mediaQuery: string, matchMedia: Function) {
   }, [mediaQuery, matchMedia]);
 
   return matches;
-}
-
-export function usePaypalButtonSetup(
-  config: any,
-  setPaypalScriptLoaded: Function,
-  paypalButtonBase?: React.FC<ButtonBaseProps>
-) {
-  /* istanbul ignore next */
-  useEffect(() => {
-    if (!config.featureFlags.usePaypalUIByDefault) {
-      return;
-    }
-
-    if (paypalButtonBase) {
-      setPaypalScriptLoaded(true);
-      return;
-    }
-
-    // Read nonce from the fxa-paypal-csp-nonce meta tag
-    const cspNonceMetaTag = document?.querySelector(
-      'meta[name="fxa-paypal-csp-nonce"]'
-    );
-    const cspNonce = JSON.parse(
-      decodeURIComponent(cspNonceMetaTag?.getAttribute('content') || '""')
-    );
-
-    const script = document.createElement('script');
-    script.src = `${config.paypal.scriptUrl}/sdk/js?client-id=${config.paypal.clientId}&vault=true&commit=false&intent=capture&disable-funding=credit,card`;
-    // Pass the csp nonce to paypal
-    script.setAttribute('data-csp-nonce', cspNonce);
-    /* istanbul ignore next */
-    script.onload = () => {
-      setPaypalScriptLoaded(true);
-    };
-    /* istanbul ignore next */
-    script.onerror = () => {
-      throw new Error('Paypal SDK could not be loaded.');
-    };
-    document.body.appendChild(script);
-  }, [config, setPaypalScriptLoaded, paypalButtonBase]);
 }
