+++ packages/fxa-payments-server/src/routes/Product/PayPalButton/index.tsx
@@ -1,9 +1,9 @@
 import React, { useCallback } from 'react';
 import ReactDOM from 'react-dom';
 
-import * as apiClient from '../../lib/apiClient';
-import { Customer } from '../../store/types';
-import { SubscriptionCreateAuthServerAPIs } from '../../routes/Product/SubscriptionCreate';
+import * as apiClient from '../../../lib/apiClient';
+import { Customer } from '../../../store/types';
+import { SubscriptionCreateAuthServerAPIs } from '../SubscriptionCreate';
 
 declare var paypal: {
   Buttons: {
@@ -12,15 +12,14 @@ declare var paypal: {
 };
 
 export type PaypalButtonProps = {
+  apiClientOverrides?: Partial<SubscriptionCreateAuthServerAPIs>;
   currencyCode: string;
   customer: Customer | null;
   idempotencyKey: string;
+  priceId: string;
   refreshSubscriptions: () => void;
   setPaymentError: Function;
-  priceId?: string;
-  newPaypalAgreement?: boolean;
-  apiClientOverrides?: Partial<SubscriptionCreateAuthServerAPIs>;
-  setTransactionInProgress?: Function;
+  setTransactionInProgress: Function;
   ButtonBase?: React.ElementType;
 };
 
@@ -31,7 +30,6 @@ export type ButtonBaseProps = {
   onClick?: (event: React.MouseEvent<HTMLButtonElement>) => void;
 };
 
-/* istanbul ignore next */
 export const PaypalButtonBase =
   typeof paypal !== 'undefined'
     ? paypal.Buttons.driver('react', {
@@ -41,14 +39,13 @@ export const PaypalButtonBase =
     : null;
 
 export const PaypalButton = ({
+  apiClientOverrides,
   currencyCode,
   customer,
   idempotencyKey,
+  priceId,
   refreshSubscriptions,
   setPaymentError,
-  priceId,
-  newPaypalAgreement,
-  apiClientOverrides,
   setTransactionInProgress,
   ButtonBase = PaypalButtonBase,
 }: PaypalButtonProps) => {
@@ -64,7 +61,6 @@ export const PaypalButton = ({
         });
       }
       const { token } = await apiGetPaypalCheckoutToken({ currencyCode });
-      /* istanbul ignore next */
       return token;
     } catch (error) {
       if (!error.code) {
@@ -83,26 +79,19 @@ export const PaypalButton = ({
 
   const onApprove = useCallback(
     async (data: { orderID: string }) => {
-      /* istanbul ignore next */
       try {
-        if (setTransactionInProgress) setTransactionInProgress(true);
-        const { apiCapturePaypalPayment, apiUpdateBillingAgreement } = {
+        setTransactionInProgress(true);
+        const { apiCapturePaypalPayment } = {
           ...apiClient,
           ...apiClientOverrides,
         };
         // This is the same token as obtained in createOrder
         const token = data.orderID;
-        if (newPaypalAgreement && priceId) {
-          await apiCapturePaypalPayment({
-            idempotencyKey,
-            priceId,
-            token,
-          });
-        } else {
-          await apiUpdateBillingAgreement({
-            token,
-          });
-        }
+        await apiCapturePaypalPayment({
+          idempotencyKey,
+          priceId,
+          token,
+        });
         refreshSubscriptions();
       } catch (error) {
         if (!error.code) {
