+++ packages/fxa-payments-server/src/routes/Subscriptions/index.tsx
@@ -284,10 +284,6 @@ export const Subscriptions = ({
           {customer.result && showPaymentUpdateForm && (
             <PaymentUpdateForm
               {...{
-                plan: planForId(
-                  customer.result.subscriptions[0].plan_id,
-                  plans.result
-                ),
                 customer: customer.result,
                 refreshSubscriptions: fetchSubscriptionsRouteResources,
                 setUpdatePaymentIsSuccess,
